+++ apps/frontend/src/pages/auth/libs/components/sign-in-form/sign-in-form.tsx
@@ -35,13 +35,13 @@ const SignInForm: React.FC<Properties> = ({ onSubmit }: Properties) => {
 
 	return (
 		<>
-			<form className={styles["form"]} onSubmit={handleFormSubmit}>
+			<form className={styles["form"]} noValidate onSubmit={handleFormSubmit}>
 				<Input
 					control={control}
 					errors={errors}
 					label="Email"
 					name="email"
-					placeholder="name@example.com"
+					placeholder="email@example.com"
 					type="email"
 				/>