+++ apps/mobile/src/libs/components/components.tsx
@@ -2,11 +2,13 @@ export { App } from "./app/app";
 export { BackgroundWrapper } from "./background-wrapper/background-wrapper";
 export { Button } from "./button/button";
 export { Checkbox } from "./checkbox/checkbox";
+export { GradientSlider } from "./gradient-slider/gradient-slider";
 export { GradientTabIcon } from "./gradient-tab-icon/gradient-tab-icon";
 export { GradientText } from "./gradient-text/gradient-text";
 export { GradientIcon } from "./icon/gradient-icon";
 export { Icon } from "./icon/icon";
 export { IconButton } from "./icon-button/icon-button";
+export { InfinitePager } from "./infinite-pager/infinite-pager";
 export { Input } from "./input/input";
 export { Link } from "./link/link";
 export { LoaderWrapper } from "./loader-wrapper/loader-wrapper";
@@ -14,10 +16,13 @@ export { Planet } from "./planet/planet";
 export { ProgressBar } from "./progress-bar/progress-bar";
 export { RadioGroup } from "./radio-group/radio-group";
 export { ScreenWrapper } from "./screen-wrapper/screen-wrapper";
+export { SliderContent } from "./slider-content/slider-content";
+export { SliderSection } from "./slider-section/slider-section";
 export { Tag } from "./tag/tag";
 export { Text } from "./text/text";
 export { ToastMessage } from "./toast-message/toast-message";
 export { Wheel } from "./wheel/wheel";
+export { WheelLoader } from "./wheel-loader/wheel-loader";
 export { default as MaskedView } from "@react-native-masked-view/masked-view";
 export {
 	ActivityIndicator,
@@ -29,10 +34,6 @@ export {
 	TouchableOpacity,
 	View,
 } from "react-native";
-export {
-	default as InfinitePager,
-	type InfinitePagerImperativeApi,
-} from "react-native-infinite-pager";
 export { LinearGradient } from "react-native-linear-gradient";
 export { default as Animated } from "react-native-reanimated";
 export {