+++ apps/mobile/src/libs/enums/app/app.ts
@@ -1,4 +1,5 @@
 export { AngleGradient } from "./angle-gradient.enum";
+export { AnimationName } from "./animation-name.enum";
 export { BaseColor } from "./base-color.enum";
 export { DataStatus } from "./data-status.enum";
 export { GradientColor } from "./gradient-color.enum";