+++ apps/mobile/src/libs/enums/navigation/root-screen-name.enum.ts
@@ -2,10 +2,12 @@ const RootScreenName = {
 	BOTTOM_TABS_NAVIGATOR: "BottomTabsNavigator",
 	CHAT: "Chat",
 	ONBOARDING: "Onboarding",
+	QUIZ: "Quiz",
 	QUIZ_ENTRY: "Quiz Entry",
 	SIGN_IN: "Sign In",
 	SIGN_UP: "Sign Up",
 	WELCOME: "Welcome",
+	WHEEL_LOADING: "Wheel Loading",
 } as const;
 
 export { RootScreenName };