+++ apps/mobile/src/libs/types/navigation/root-navigation-parameter-list.type.ts
@@ -4,10 +4,12 @@ type RootNavigationParameterList = {
 	[RootScreenName.BOTTOM_TABS_NAVIGATOR]: undefined;
 	[RootScreenName.CHAT]: undefined;
 	[RootScreenName.ONBOARDING]: undefined;
+	[RootScreenName.QUIZ]: undefined;
 	[RootScreenName.QUIZ_ENTRY]: undefined;
 	[RootScreenName.SIGN_IN]: undefined;
 	[RootScreenName.SIGN_UP]: undefined;
 	[RootScreenName.WELCOME]: undefined;
+	[RootScreenName.WHEEL_LOADING]: undefined;
 };
 
 export { type RootNavigationParameterList };