+++ apps/mobile/src/libs/types/types.ts
@@ -1,5 +1,6 @@
 export { type AsyncThunkConfig } from "./app/app";
 export { type IconName } from "./icon/icon";
+export { type InfinitePagerImperativeApi } from "./infinite-pager/infinite-pager";
 export {
 	type BottomTabNavigationParameterList,
 	type NavigationScreenProperties,
@@ -9,7 +10,13 @@ export {
 	type RadioGroupItemStyle,
 	type RadioGroupOption,
 } from "./radio-group/types";
-export { type WheelDataItem } from "./wheel/wheel";
+export { type SliderData } from "./slider/slider";
+export {
+	type SectorInitialSharedValues,
+	type SectorParameters,
+	type SectorParametersCalculationData,
+	type WheelDataItem,
+} from "./wheel/wheel";
 export { type NativeStackNavigationProp } from "@react-navigation/native-stack";
 export {
 	type Control,
@@ -19,12 +26,14 @@ export {
 } from "react-hook-form";
 export {
 	type ImageSourcePropType,
+	type LayoutChangeEvent,
 	type StyleProp,
 	type TextStyle,
 	type ViewStyle,
 } from "react-native";
 export { type LinearGradientProps } from "react-native-linear-gradient";
 export { type RadioButtonProps } from "react-native-radio-buttons-group";
+export { type SharedValue } from "react-native-reanimated";
 export {
 	type OnboardingQuestionResponseDto,
 	type ServerErrorDetail,