+++ apps/mobile/src/navigations/root/root.tsx
@@ -10,8 +10,10 @@ import { type RootNavigationParameterList } from "~/libs/types/types";
 import { BottomTabsNavigator } from "~/navigations/bottom-tabs-navigator/bottom-tabs-navigator";
 import { Auth } from "~/screens/auth/auth";
 import { Onboarding } from "~/screens/onboarding/onboarding";
+import { Quiz } from "~/screens/quiz/quiz";
 import { QuizEntry } from "~/screens/quiz-entry/quiz-entry";
 import { Welcome } from "~/screens/welcome/welcome";
+import { WheelLoading } from "~/screens/wheel-loading/wheel-loading";
 
 const NativeStack = createNativeStackNavigator<RootNavigationParameterList>();
 
@@ -38,6 +40,11 @@ const Root: React.FC = () => {
 					<NativeStack.Screen
 						component={QuizEntry}
 						name={RootScreenName.QUIZ_ENTRY}
+					/>
+					<NativeStack.Screen component={Quiz} name={RootScreenName.QUIZ} />
+					<NativeStack.Screen
+						component={WheelLoading}
+						name={RootScreenName.WHEEL_LOADING}
 					/>
 					<NativeStack.Screen
 						component={BottomTabsNavigator}